# Conflicts:
#	src/main/java/com/gym/crm/app/security/AuthenticationService.java
#	src/test/java/com/gym/crm/app/security/AuthenticationServiceTest.java

+++ src/main/java/com/gym/crm/app/facade/GymFacade.java
@@ -8,14 +8,16 @@ import com.gym.crm.app.domain.dto.trainer.TrainerDto;
 import com.gym.crm.app.domain.dto.trainer.TrainerUpdateRequest;
 import com.gym.crm.app.domain.dto.training.TrainingDto;
 import com.gym.crm.app.domain.dto.training.TrainingSaveRequest;
-import com.gym.crm.app.domain.dto.user.UserCredentialsDto;
+import com.gym.crm.app.domain.model.TrainingType;
 import com.gym.crm.app.mapper.TraineeMapper;
 import com.gym.crm.app.mapper.TrainerMapper;
 import com.gym.crm.app.mapper.TrainingMapper;
+import com.gym.crm.app.mapper.TrainingTypeMapper;
 import com.gym.crm.app.mapper.UserMapper;
 import com.gym.crm.app.repository.search.filters.TraineeTrainingSearchFilter;
 import com.gym.crm.app.repository.search.filters.TrainerTrainingSearchFilter;
 import com.gym.crm.app.rest.AvailableTrainerGetResponse;
+import com.gym.crm.app.rest.ChangePasswordRequest;
 import com.gym.crm.app.rest.TraineeAssignedTrainersUpdateRequest;
 import com.gym.crm.app.rest.TraineeAssignedTrainersUpdateResponse;
 import com.gym.crm.app.rest.TraineeCreateResponse;
@@ -27,20 +29,29 @@ import com.gym.crm.app.rest.TrainerCreateResponse;
 import com.gym.crm.app.rest.TrainerGetResponse;
 import com.gym.crm.app.rest.TrainerTrainingGetResponse;
 import com.gym.crm.app.rest.TrainerUpdateResponse;
+import com.gym.crm.app.rest.TrainingCreateRequest;
+import com.gym.crm.app.rest.TrainingTypeGetResponse;
 import com.gym.crm.app.rest.TrainingWithTraineeName;
 import com.gym.crm.app.rest.TrainingWithTrainerName;
 import com.gym.crm.app.security.AuthenticationService;
 import com.gym.crm.app.security.CurrentUserHolder;
+import com.gym.crm.app.security.model.UserCredentialsDto;
 import com.gym.crm.app.service.TraineeService;
 import com.gym.crm.app.service.TrainerService;
 import com.gym.crm.app.service.TrainingService;
 import com.gym.crm.app.service.common.UserProfileService;
 import jakarta.validation.Valid;
+import jakarta.validation.constraints.NotNull;
 import lombok.Getter;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Component;
 
+import java.math.BigDecimal;
 import java.util.List;
+
+import static com.gym.crm.app.security.UserRole.ADMIN;
+import static com.gym.crm.app.security.UserRole.TRAINEE;
+import static com.gym.crm.app.security.UserRole.TRAINER;
 
 @Getter
 @Component
@@ -53,6 +64,7 @@ public class GymFacade {
     private final TraineeMapper traineeMapper;
     private final TrainerMapper trainerMapper;
     private final TrainingMapper trainingMapper;
+    private final TrainingTypeMapper trainingTypeMapper;
 
     private final UserMapper userMapper;
     private final CurrentUserHolder currentUserHolder;
@@ -64,7 +76,7 @@ public class GymFacade {
                      UserProfileService userProfileService,
                      AuthenticationService authService,
                      TraineeMapper traineeMapper,
-                     TrainerMapper trainerMapper, TrainingMapper trainingMapper,
+                     TrainerMapper trainerMapper, TrainingMapper trainingMapper, TrainingTypeMapper trainingTypeMapper,
                      UserMapper userMapper, CurrentUserHolder currentUserHolder) {
         this.traineeService = traineeService;
         this.trainerService = trainerService;
@@ -74,6 +86,7 @@ public class GymFacade {
         this.traineeMapper = traineeMapper;
         this.trainerMapper = trainerMapper;
         this.trainingMapper = trainingMapper;
+        this.trainingTypeMapper = trainingTypeMapper;
         this.userMapper = userMapper;
         this.currentUserHolder = currentUserHolder;
     }
@@ -89,32 +102,32 @@ public class GymFacade {
     }
 
     public List<TrainerDto> getAllTrainers(UserCredentialsDto userCredentials) {
-        authService.authenticate(userCredentials);
+        authService.checkUserAuthorisation(userCredentials, ADMIN);
 
         return trainerService.getAllTrainers();
     }
 
     public List<TraineeDto> getAllTrainees(UserCredentialsDto userCredentials) {
-        authService.authenticate(userCredentials);
+        authService.checkUserAuthorisation(userCredentials, ADMIN);
 
         return traineeService.getAllTrainees();
     }
 
     public TrainerGetResponse getTrainerByUsername(String username) {
-        authService.authenticate(getCurrentCredentials());
+        authService.checkUserAuthorisation(getCurrentCredentials(), ADMIN, TRAINER, TRAINEE);
 
         return trainerMapper.toGetResponse(trainerService.getTrainerByUsername(username));
     }
 
     public TraineeGetResponse getTraineeByUsername(String username) {
-        authService.authenticate(getCurrentCredentials());
+        authService.checkUserAuthorisation(getCurrentCredentials(), ADMIN, TRAINER, TRAINEE);
 
         return traineeMapper.dtoToGetResponse(traineeService.getTraineeByUsername(username));
     }
 
     public TraineeAssignedTrainersUpdateResponse updateTraineeTrainersList(String username,
                                                                            TraineeAssignedTrainersUpdateRequest request) {
-        authService.authenticate(getCurrentCredentials());
+        authService.checkUserAuthorisation(getCurrentCredentials(), ADMIN, TRAINER, TRAINEE);
 
         List<Trainer> trainers = traineeService.updateTraineeTrainersByUsername(username, request.getTrainerUsernames()).stream()
                 .map(trainerMapper::entityToRestTrainer).toList();
@@ -124,36 +137,36 @@ public class GymFacade {
 
     public TraineeUpdateResponse updateTraineeByUsername(String username,
                                                          @Valid TraineeUpdateRequest updateRequest) {
-        authService.authenticate(getCurrentCredentials());
+        authService.checkUserAuthorisation(getCurrentCredentials(), ADMIN, TRAINEE);
 
         return traineeMapper.dtoToUpdateResponse(traineeService.updateTraineeByUsername(username, updateRequest));
     }
 
     public TrainerUpdateResponse updateTrainerByUsername(String username,
                                                          @Valid TrainerUpdateRequest updateRequest) {
-        authService.authenticate(getCurrentCredentials());
+        authService.checkUserAuthorisation(getCurrentCredentials(), ADMIN, TRAINER);
 
         return trainerMapper.toUpdateResponse(trainerService.updateTrainerByUsername(username, updateRequest));
     }
 
     public void deleteTrainerByUsername(String username, UserCredentialsDto userCredentials) {
-        authService.authenticate(userCredentials);
+        authService.checkUserAuthorisation(userCredentials, ADMIN, TRAINER);
 
         trainerService.deleteTrainerByUsername(username);
     }
 
     public void deleteTraineeByUsername(String username) {
-        authService.authenticate(getCurrentCredentials());
+        authService.checkUserAuthorisation(getCurrentCredentials(), ADMIN, TRAINEE);
 
         traineeService.deleteTraineeByUsername(username);
     }
 
     public TrainerTrainingGetResponse getTrainerTrainingsByFilter(@Valid TrainerTrainingSearchFilter criteria) {
-        authService.authenticate(getCurrentCredentials());
+        authService.checkUserAuthorisation(getCurrentCredentials(), ADMIN, TRAINER);
 
         List<TrainingDto> trainings = trainingService.getTrainerTrainingsByFilter(criteria);
 
-        List<TrainingWithTraineeName>trainingWithTraineeNames = trainings.stream()
+        List<TrainingWithTraineeName> trainingWithTraineeNames = trainings.stream()
                 .map(this::buildTrainingWithTraineeName)
                 .toList();
 
@@ -161,7 +174,7 @@ public class GymFacade {
     }
 
     public TraineeTrainingGetResponse getTraineeTrainingsByFilter(@Valid TraineeTrainingSearchFilter filter) {
-        authService.authenticate(getCurrentCredentials());
+        authService.checkUserAuthorisation(getCurrentCredentials(), ADMIN, TRAINER, TRAINEE);
 
         List<TrainingDto> trainings = trainingService.getTraineeTrainingsByFilter(filter);
 
@@ -173,7 +186,7 @@ public class GymFacade {
     }
 
     public AvailableTrainerGetResponse getUnassignedTrainersByTraineeUsername(String username) {
-        authService.authenticate(getCurrentCredentials());
+        authService.checkUserAuthorisation(getCurrentCredentials(), ADMIN, TRAINEE);
 
         List<Trainer> trainers = traineeService.getUnassignedTrainersByTraineeUsername(username).stream()
                 .map(trainerMapper::toEntity).toList();
@@ -182,38 +195,56 @@ public class GymFacade {
     }
 
     public List<TrainingDto> getAllTrainings(UserCredentialsDto userCredentials) {
-        authService.authenticate(userCredentials);
+        authService.checkUserAuthorisation(userCredentials, ADMIN);
 
         return trainingService.getAllTrainings();
     }
 
-    public TrainingDto addTraining(@Valid TrainingSaveRequest createRequest, UserCredentialsDto userCredentials) {
-        authService.authenticate(userCredentials);
+    public TrainingDto addTraining(@Valid TrainingCreateRequest request) {
+        authService.checkUserAuthorisation(getCurrentCredentials(), ADMIN);
 
-        return trainingService.addTraining(createRequest);
+        TrainerDto trainer = trainerService.getTrainerByUsername(request.getTrainerUsername());
+
+        TrainingSaveRequest saveRequest = new TrainingSaveRequest();
+        saveRequest.setTrainingName(request.getTrainingName());
+        saveRequest.setTrainingDate(request.getTrainingDate());
+        saveRequest.setTrainingDuration(BigDecimal.valueOf(request.getTrainingDuration()));
+        saveRequest.setTrainingTypeName(trainer.getSpecialization().getTrainingTypeName());
+        saveRequest.setTraineeId(traineeService.getTraineeByUsername(request.getTraineeUsername()).getUserId());
+        saveRequest.setTrainerId(trainer.getUserId());
+
+        return trainingService.addTraining(saveRequest);
+    }
+
+    public TrainingTypeGetResponse getAllTrainingsTypes() {
+        authService.checkUserAuthorisation(getCurrentCredentials(), ADMIN, TRAINEE, TRAINER);
+
+        List<TrainingType> trainingTypes = trainingService.getTrainingTypes();
+        var trainingTypesRest = trainingTypes.stream().map(trainingTypeMapper::toRestTrainingType).toList();
+
+        return new TrainingTypeGetResponse().trainingTypes(trainingTypesRest);
     }
 
     public TrainingDto updateTraining(@Valid TrainingSaveRequest updateRequest, UserCredentialsDto userCredentials) {
-        authService.authenticate(userCredentials);
+        authService.checkUserAuthorisation(userCredentials, ADMIN);
 
         return trainingService.updateTraining(updateRequest);
     }
 
     public void switchActivationStatus(String username) {
-        authService.authenticate(getCurrentCredentials());
+        authService.checkUserAuthorisation(getCurrentCredentials(), ADMIN, TRAINER, TRAINEE);
 
         userProfileService.switchActivationStatus(username);
     }
 
-    public void changePassword(String username, String password, UserCredentialsDto userCredentials) {
-        authService.authenticate(userCredentials);
+    public void changePassword(ChangePasswordRequest request) {
+        authService.checkUserAuthorisation(getCurrentCredentials(), ADMIN, TRAINER, TRAINEE);
 
-        userProfileService.changePassword(username, password);
+        userProfileService.changePassword(request.getUsername(), request.getOldPassword(), request.getNewPassword());
     }
 
     private UserCredentialsDto getCurrentCredentials() {
-        return userMapper.toCredentialsDto(
-                currentUserHolder.get());
+        return userMapper.toCredentialsDto(currentUserHolder.get());
     }
 
     private TrainingWithTrainerName buildTrainingWithTrainerName(TrainingDto trainingDto) {
@@ -243,5 +274,4 @@ public class GymFacade {
 
         return trainingWithTraineeName;
     }
-
 }
